+++ src/components/Card/Card.css
@@ -2,16 +2,13 @@
 	width: 86px;
 	height: 120px;
 	border-radius: 8px;
-	object-fit: cover;
+	object-fit: contain;
 	user-select: none;
 	display: block;
-	aspect-ratio: 43/60;
-	background: #fff;
 }
 .card-img.small {
 	width: 64px;
 	height: 90px;
-	aspect-ratio: 32/45;
 }
 .card-skel { width: 86px; height: 120px; background: #072113; border-radius: 8px; border: 1px dashed rgba(255,255,255,0.06); }
 .card-skel.small { width: 64px; height: 90px; }
