+++ src/components/DealerArea/DealerArea.css
@@ -16,3 +16,28 @@
   flex-direction: column;
   align-items: center;
 }
+
+.dealer-hand {
+  background: rgba(0,0,0,0.24);
+  padding: 10px;
+  border-radius: 12px;
+  min-width: 220px;
+  text-align: center;
+  border: 1px solid rgba(255,255,255,0.04);
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  box-shadow: 0 6px 18px rgba(0,0,0,0.35);
+  border-color: rgba(255,255,255,0.12);
+}
+
+.dealer.cards,
+.dealer-cards {
+  display: flex;
+  flex-direction: row;
+  gap: 8px;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: nowrap;
+  overflow-x: auto;
+}
